SPAKE2_Symmetric: switch to faster M=N version
This sets the two blinding factors equal to each other (but still a
group element for which we don't know the discrete log, of course). This
reduces the message size and processing speed in half (making it equal
to the asymmetric case). I believe this to be safe, based upon a proof
by Kobara/Imai in 2003 and discussions with Mike Hamburg and other folks
since then (README.md has the references), but the asymmetric form is
still there for folks who aren't comfortable with that.