move parameters into separately importable modules, for speed
Computing the "arbitrary elements" for the integer groups was adding 400ms to
magic-wormhole's import time on my laptop, which is a lot for groups that it
doesn't use.

This is a compatibility break for any application that uses non-default
parameters (i.e. "from spake2 import Params1024"). Such apps need to be
changed to use "from spake2.parameters.i1024 import Params1024".